Port to Acorn 6 plugin API
